pyproject.toml
[project]
name = "moneymitra-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "langchain[google-genai]>=1.2.15",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
